v1/05-er-triage-medpalm-demo/requirements.txt
langchain
streamlit
google-cloud-aiplatform
black
python-dotenv
google-generativeai
pinecone-client
google-api-python-client
google-auth-oauthlib